arm/x86/vmap: Add vmalloc_xen and vm_init_type
For those users who want to use the virtual addresses that
are in the hypervisor's code/data region address space -
these three new functions allow that.

Implementation wise the vmap API keeps track of two virtual
address regions now:
 a) VMAP_VIRT_START
 b) Any provided virtual address space (need start and end).

The a) one is the default one and the existing behavior
for users of vmalloc, vmap, etc is the same.

If however one wishes to use the b) one only has to use
the vm_init_type to initialize and the vmzalloc_xen to utilize
it (vfree and vunmap are capable of searching both address spaces).

This allows users (such as xSplice) to provide their own
mechanism to change the the page flags, and also use virtual
addresses closer to the hypervisor virtual addresses (at least
on x86) while not having to deal with the allocation of
pages.

For example of users, see patch titled "xsplice: Implement payload
loading", where we parse the payload's ELF relocations - which
is defined to be signed 32-bit (on x86) (max displacement hence
is 2GB virtual space, ARM32 is 128MB). The displacement of the
hypervisor virtual addresses to the vmalloc (on x86)
is more than 32-bits - which means that ELF relocations would
truncate the 34 and 33th bit. Hence this alternate API.

We also add add extra checks in case the b) range has not been
initialized.

Part of this patch also removes 'vm_alloc' and 'vm_free'
decleration as we do not have any users of it.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Suggested-by: Jan Beulich <jbeulich@suse.com>
Acked-by: Julien Grall <julien.grall@arm.com> [ARM]
Reviewed-by: Jan Beulich <jbeulich@suse.com>
Release-acked-by: Wei Liu <wei.liu2@citrix.com>
(cherry picked from commit 9e3e7117770ace6e0cb2a50dfc4859a3d9ce7b6b)
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>

 Conflicts:
	xen/arch/arm/kernel.c
	xen/arch/arm/mm.c
[We don't have all the ARM backports]
	xen/arch/x86/mm.c
	xen/common/vmap.c
[B/c we did not backport c2453291b "xen: Add populate_pt_range interface
 to reserve non-leaf level table entries" - on x86 the populate_pt_range
 is just a wrapper around map_pages_to_xen]
	xen/drivers/acpi/osl.c