livepatch: NOP if func->new_size is zero. WIP
The NOP functionality will NOP any of the code at
the 'old_addr' or at 'name'. Since the purpose of this
is to NOP out calls, such as:

 e9 <4-bytes-offset>

(5 byte insn), or on ARM 4 byte insn for branching
- we assume that the NOP is only for platform specific
call sites.

As such if the `old_addr` is provided we will NOP
5 instructions (on x86) at that location.

If `name` is provided with the symbol+0x<offset/<len>
we make sure that <len> is 5.

While at it, also unify the code on x86 doing the
patching to be a bit simpler.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>