livepatch: Add parsing for the symbol+0x<offset>/<len>
in case we want to patch at specific offsets inside
a function. (for example if we want to do NOP patching).

We also assume that the 'len' is only the size of
an isns that would be for a call opcode (so 5 bytes
on x86, and 4 on ARM 32/64).

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
---
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: Ross Lagerwall <ross.lagerwall@citrix.com>
Cc: Stefano Stabellini <sstabellini@kernel.org>
Cc: Julien Grall <julien.grall@arm.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>

v3: First submission