livepatch: ARM64 support alternative patching.
With the ALT_[ORIG|REPL]_PTR macro being the same on x86 and
ARM64 we can wrap access to the structure using this macro.

Unfortunatly the call to patch is behind a different
a function so we still have to keep an #ifdef.

The #ifndef CONFIG_ARM is moved as exceptions are not part
of ARM64. But unfortunatly we still have to guard the
code as ARM32 does not support alternative patching.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>