livepatch: ARM32 support. WIP
First of, it does not work. After patching the call-site (and verying
it has been patched with the right instruction), we end up with:

CPU0: Unexpected Trap: Prefetch Abort

whenever the new function is called.

Otherwise it works :-)

The ELF relocation was the most hardest as the initial addendum
is coded within the instruction - hence before we do any relocation
we have to extract the addendum.

Xen hypervisor, when compiled only emits ELF REL structures, but
in case we do have ELF RELA structures - that component has also
been implement.

The patch piggybacks on: livepatch: Initial ARM64 support, which
brings up all of the neccessary livepatch infrastructure pieces in.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
---
Cc: Julien Grall <julien.grall@arm.com>
Cc: Stefano Stabellini <sstabellini@kernel.org>

v2: First submission.