livepatch/tests: Move the .name value to .rodata
Right now the contents of 'name' are all located in
the .data section. We want them in the .rodata section
so change the type to have const on them.

Reviewed-by: Ross Lagerwall <ross.lagerwall@citrix.com>
Acked-by: Jan Beulich <jbeulich@suse.com>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
---
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: Jan Beulich <jbeulich@suse.com>

v3: First submission (as part of the ARM 32 and 64 support of Livepatch)
v6: Now moved to the Common Livepatch fixes for 4.8
v7: Add Jan's Ack.
    Added Ross's Reviewed-by.