tmem: Untangle dedup, compression and trailing zero options.
One can enable _all_ of them and then things mostly work. Albeit
if one looks carefully in the hypervisor ring buffer the
tze is disabled if compression (and dedup) is enabled.

However if you look in the handline of XEN_SYSCTL_TMEM_OP_SET_COMPRESS
it says:
  if ( tmem_dedup_enabled() )
        ("tmem: compression %s for all %ss, cannot be changed when tmem_dedup is enabled\n",

which implies that you cannot have dedup and compression enabled
either!

So putting that as well so that if somebody enables
_all_ of them, only dedup is enabled.

Also while at it - move the printk telling user what is
enabled to after the various knobs have been turned off.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>